feat: Restructure e2e testing
